backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Windows integration
pywin32==306; sys_platform == "win32"
pywinrm==0.4.3; sys_platform == "win32"

# Data processing
pandas==2.1.3
numpy==1.24.4
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# AI/ML
requests==2.31.0
aiohttp==3.9.0
httpx==0.25.2

# Task queue
celery[redis]==5.3.4
flower==2.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# File handling
python-magic==0.4.27
Pillow==10.1.0
aiofiles==23.2.1

# Validation
pydantic[email]==2.5.0
validators==0.22.0

# Logging
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1